Some additions to build system

+++ src/conv.c
@@ -1,22 +1,23 @@
 // Split program size in half, since 16 instructions
+// This not actually used, just a demo.
 
 #include <stdio.h>
 #include <stdlib.h>
 
 char inst[] = "!%*+-<>ad^v.,$?|";
 
-size_t getChar(char a) {
-    size_t c;
+int getChar(char a) {
+    int c;
     for (c = 0; inst[c] != a && inst[c] != '\0'; c++);
     return c;
 }
 
 int main(int argc, char *argv[]) {
-    for (size_t i = 0; argv[1][i] != '\0'; i++) {
-        size_t a = getChar(argv[1][i]);
+    for (int i = 0; argv[1][i] != '\0'; i++) {
+        int a = getChar(argv[1][i]);
         i++;
-        size_t b = getChar(argv[1][i]);
-        printf("%lu, ", (a << 4) | b);
+        int b = getChar(argv[1][i]);
+        printf("%d, ", (a << 4) | b);
     }
 
     return 0;
